delete: AbstractSyntaxTree and the interpreter of that implementation
