+++ ast/src/main/kotlin/ast/AST.kt
@@ -218,42 +218,3 @@ class AssignmentStatement(override val range: Range, val variableName: String, v
         return result
     }
 }
-
-/**
- * Proposal of AST, by Miguel
- * Need add Range to Nodes data
- */
-
-enum class PrintScriptType {
-    STRING,
-    NUMBER,
-}
-
-// TODO: merge both proposals, bringing out the best of each one.
-data class AbstractSyntaxTree(val nodes: List<Node>)
-
-abstract class Node
-
-abstract class ExpressionNode : Node()
-
-class LiteralNode(val type: PrintScriptType, val value: String) : ExpressionNode()
-
-class IdentifierNode(val identifier: String) : ExpressionNode()
-
-class AdditionNode(val left: ExpressionNode, val right: ExpressionNode) : ExpressionNode()
-
-abstract class SimpleBinaryOperation(val left: ExpressionNode, val right: ExpressionNode) : ExpressionNode()
-
-class SubtractionNode(left: ExpressionNode, right: ExpressionNode) : SimpleBinaryOperation(left, right)
-
-class MultiplicationNode(left: ExpressionNode, right: ExpressionNode) : SimpleBinaryOperation(left, right)
-
-class DivisionNode(left: ExpressionNode, right: ExpressionNode) : SimpleBinaryOperation(left, right)
-
-class DeclareNode(val type: PrintScriptType, val identifier: String) : Node()
-
-class AssignNode(val identifier: String, val value: ExpressionNode) : Node()
-
-class DeclareAndAssignNode(val declare: DeclareNode, val value: ExpressionNode) : Node()
-
-class PrintNode(val argument: ExpressionNode) : Node()
